websocket-server/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-jose==3.3.0
pydantic==2.9.2
